Fire desktop notifications from the daemon when the TUI is parked or closed
Desktop popups were fired only by the TUI's detect_notifications, which runs in
its event loop. When you "go all the way in" (attach to an agent's tmux pane),
that loop parks on the blocking tmux attach — so no edge detection and no popups,
for any agent, until you detach. Same when the TUI is closed entirely.

Make the always-running daemon cover the gap:
- Lift the local desktop delivery (send_native + osascript/terminal-notifier
  helpers) into repomon_core::notify so both the TUI and the daemon can fire it.
- The local socket stamps a "TUI seen" time on every request; the TUI's ~1s
  lane.list refresh is a built-in heartbeat that stops the instant it parks or
  closes. Remote/WS requests don't stamp it, so iOS polling can't mask a parked
  local TUI.
- notify_watch now runs whenever notify_enabled (not only with the remote bridge)
  and fires send_native for each transition only when the TUI heartbeat is stale,
  so there's never a double-notify while the TUI is actively watching. Remote
  event.notification + APNs stay gated on the bridge.

Net: an agent going "needs you" reaches you on the desktop whether the TUI is
open, attached full-screen, or closed. Idle CPU unchanged (~0.4%).
